+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.7] - 2023-09-21
+
+### Added
+- Added `@typdefs` for person, vital, measurement, record
+
+### Changed
+- Measurement Value now saves as number
+- Vital High and Low values now save as number
+- Made views, stores and routes more CRUDy
+
+### Fixed
+- Fixed routing issue with creating and updating a person
+
+
 ## [0.1.6] - 2023-09-20
 
 ### Added